split stats.rs into submodules
as requested by @gjcolombo
