+++ bin/propolis-server/src/lib/stats.rs
@@ -14,7 +14,6 @@ use oximeter::{
     Metric, MetricsError, Producer, Target,
 };
 use oximeter_producer::{Config, Server};
-use propolis::hw::qemu::pvpanic::QemuPvpanic;
 use slog::{error, info, warn, Logger};
 
 use std::net::SocketAddr;
@@ -22,6 +21,9 @@ use std::sync::{Arc, Mutex};
 use uuid::Uuid;
 
 use crate::server::MetricsEndpointConfig;
+
+mod pvpanic;
+pub use self::pvpanic::PvpanicProducer;
 
 const OXIMETER_STAT_INTERVAL: tokio::time::Duration =
     tokio::time::Duration::from_secs(30);
@@ -39,24 +41,6 @@ struct InstanceUuid {
 #[derive(Debug, Default, Copy, Clone, Metric)]
 struct Reset {
     /// The number of times this instance was reset via the API.
-    #[datum]
-    pub count: Cumulative<i64>,
-}
-
-/// An Oximeter `Metric` that specifies the number of times an instance's guest
-/// reported a guest-handled kernel panic using the QEMU `pvpanic` device.
-#[derive(Debug, Default, Copy, Clone, Metric)]
-struct PvPanicGuestHandled {
-    /// The number of times this instance's guest handled a kernel panic.
-    #[datum]
-    pub count: Cumulative<i64>,
-}
-
-/// An Oximeter `Metric` that specifies the number of times an instance's guest
-/// reported a host-handled kernel panic using the QEMU `pvpanic` device.
-#[derive(Debug, Default, Copy, Clone, Metric)]
-struct PvPanicHostHandled {
-    /// The number of times this instance's reported a host-handled kernel panic.
     #[datum]
     pub count: Cumulative<i64>,
 }
@@ -89,19 +73,6 @@ pub struct ServerStatsOuter {
     server_stats_wrapped: Arc<Mutex<ServerStats>>,
 }
 
-#[derive(Clone, Debug)]
-pub struct PvpanicProducer {
-    /// The name to use as the Oximeter target, i.e. the identifier of the
-    /// source of these metrics.
-    stat_name: InstanceUuid,
-
-    /// Kernel panic counts for the relevant instance.
-    host_handled_panics: PvPanicHostHandled,
-    guest_handled_panics: PvPanicGuestHandled,
-
-    counts: Arc<QemuPvpanic>,
-}
-
 impl ServerStatsOuter {
     /// Increments the number of times the instance was reset.
     pub fn count_reset(&self) {
@@ -118,36 +89,6 @@ impl Producer for ServerStatsOuter {
         let inner = self.server_stats_wrapped.lock().unwrap();
         let name = inner.stat_name;
         let data = vec![Sample::new(&name, &inner.run_count)?];
-        Ok(Box::new(data.into_iter()))
-    }
-}
-
-impl PvpanicProducer {
-    pub fn new(id: Uuid, counts: Arc<QemuPvpanic>) -> Self {
-        PvpanicProducer {
-            stat_name: InstanceUuid { uuid: id },
-            host_handled_panics: Default::default(),
-            guest_handled_panics: Default::default(),
-            counts,
-        }
-    }
-}
-
-impl Producer for PvpanicProducer {
-    fn produce(
-        &mut self,
-    ) -> Result<Box<dyn Iterator<Item = Sample> + 'static>, MetricsError> {
-        self.host_handled_panics
-            .datum_mut()
-            .set(self.counts.host_handled_count() as i64);
-        self.guest_handled_panics
-            .datum_mut()
-            .set(self.counts.guest_handled_count() as i64);
-        let data = vec![
-            Sample::new(&self.stat_name, &self.guest_handled_panics)?,
-            Sample::new(&self.stat_name, &self.host_handled_panics)?,
-        ];
-
         Ok(Box::new(data.into_iter()))
     }
 }
